requirements.txt
flask
Flask-security
Flask-login
Flask-mailman
Flask-principal
Flask-WTF
itsdangerous
passlib
QRCode
webauthn
jinja2
python-decouple
email_validator
flask_bcrypt
flask_sqlalchemy
sqlalchemy
flask_migrate
pyopenssl
stripe
blinker
Click
itsdangerous
MarkupSafe
mysql-connector-python
mysql-connector-python-rf
mysql-connector